requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OAuth/OIDC
authlib>=1.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0

# Security
cryptography>=41.0.0
itsdangerous>=2.1.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.2

# Biometric (optional)
# webauthn>=1.0.0
